refactor(app): extract LSP change-notification orchestrators (phase 3h)
Move collect_lsp_changes, calculate_event_line_info, notify_lsp_save,
notify_lsp_save_buffer out of render.rs into a new
app/lsp_event_notify.rs sibling. ~237 lines.

These translate Editor `Event`s into LSP
`TextDocumentContentChangeEvent`s, compute line-shift metadata for
plugin hooks, and send `did_save` notifications to running LSP servers.

render.rs goes from 2136 -> 1894 lines (-242). 2026 tests still pass.
